+++ app/products/models/__init__.py
@@ -1,0 +1,5 @@
+from .intermediate import *
+from .manytomany import *
+from .manytoone import *
+from .onetoone import *
+from .product import *
